+++ src/main/java/by/epamtc/digapply/mapper/UserRowMapper.java
@@ -12,12 +12,12 @@ public class UserRowMapper implements RowMapper<User> {
     @Override
     public User map(ResultSet resultSet) throws SQLException {
         User user = new User();
-        user.setUserId(resultSet.getInt("user_id"));
+        user.setUserId(resultSet.getLong("user_id"));
         user.setEmail(resultSet.getString("email"));
         user.setPassword(resultSet.getString("password"));
         user.setName(resultSet.getString("name"));
         user.setSurname(resultSet.getString("surname"));
-        user.setRoleId(resultSet.getInt("role_id"));
+        user.setRoleId(resultSet.getLong("role_id"));
         return user;
     }
 }
